+++ lib/lottie/objects/layers.py
@@ -4,8 +4,9 @@ from .effects import Effect
 from .helpers import Transform, Mask, VisualObject, BlendMode
 from .shapes import ShapeElement
 from .text import TextAnimatorData
-from .properties import Value
+from .properties import Value, MultiDimensional
 from ..utils.color import Color, color_from_hex, color_to_hex
+from .styles import LayerStyle
 
 
 ## @ingroup Lottie
@@ -17,39 +18,86 @@ class MatteMode(LottieEnum):
     Luma = 3
     InvertedLuma = 4
 
-
-## @ingroup Lottie
+#ingroup Lottie
 class Layer(VisualObject):
+    """!
+    Base class for all layers
+    """
     _props = [
         LottieProp("threedimensional", "ddd", PseudoBool, False),
         LottieProp("hidden", "hd", bool, False),
         LottieProp("type", "ty", int, False),
+        LottieProp("index", "ind", int, False),
         LottieProp("parent_index", "parent", int, False),
-
-        LottieProp("stretch", "sr", float, False),
-        LottieProp("transform", "ks", Transform, False),
-        LottieProp("auto_orient", "ao", PseudoBool, False),
-
+        LottieProp("time_stretch", "sr", float, False),
         LottieProp("in_point", "ip", float, False),
         LottieProp("out_point", "op", float, False),
         LottieProp("start_time", "st", float, False),
+    ]
+    ## %Layer type.
+    ## @see https://github.com/bodymovin/bodymovin-extension/blob/master/bundle/jsx/enums/layerTypes.jsx
+    type = None
+    _classses = {}
+
+    def __init__(self):
+        super().__init__()
+
+        ## Whether the layer is threedimensional
+        self.threedimensional = 0
+        ## Whether the layer is hidden
+        self.hidden = None
+        ## Index that can be used for parenting and referenced in expressions
+        self.index = None
+        ## Must be the `ind` property of another layer
+        self.parent_index = None
+        self.time_stretch = 1
+        ## Frame when the layer becomes visible
+        self.in_point = None
+        ## Frame when the layer becomes invisible
+        self.out_point = None
+        ## Start Time of layer. Sets the start time of the layer.
+        self.start_time = 0
+
+    @classmethod
+    def _load_get_class(cls, lottiedict):
+        if not Layer._classses:
+            Layer._classses = {
+                sc.type: sc
+                for sc in Layer.__subclasses__() + VisualLayer.__subclasses__()
+                if sc.type is not None
+            }
+        type_id = lottiedict["ty"]
+        if type_id not in Layer._classses:
+            warnings.warn("Unknown layer type: %s" % type_id)
+            return Layer
+        return Layer._classses[type_id]
+
+
+## @ingroup Lottie
+class VisualLayer(Layer):
+    """!
+    Base class for layers that have a visual component
+    """
+    _props = [
+        LottieProp("collapse_transform", "cp", bool, False),
+        LottieProp("transform", "ks", Transform, False),
+        LottieProp("auto_orient", "ao", PseudoBool, False),
+
         LottieProp("blend_mode", "bm", BlendMode, False),
 
         LottieProp("matte_mode", "tt", MatteMode, False),
-        LottieProp("index", "ind", int, False),
         LottieProp("css_class", "cl", str, False),
         LottieProp("layer_xml_id", "ln", str, False),
+        LottieProp("layer_xml_tag_name", "tg", str, False),
+
         LottieProp("motion_blur", "mb", bool, False),
+        LottieProp("layer_style", "sy", LayerStyle, True),
 
         LottieProp("has_masks", "hasMask", bool, False),
         LottieProp("masks", "masksProperties", Mask, True),
         LottieProp("effects", "ef", Effect, True),
         LottieProp("matte_target", "td", int, False),
     ]
-    ## %Layer type.
-    ## @see https://github.com/bodymovin/bodymovin-extension/blob/master/bundle/jsx/enums/layerTypes.jsx
-    type = None
-    _classses = {}
 
     @property
     def has_masks(self):
@@ -60,36 +108,29 @@ class Layer(VisualObject):
 
     def __init__(self):
         super().__init__()
+
+        self.collapse_transform = None
         ## Transform properties
         self.transform = Transform()
         ## Auto-Orient along path AE property.
         self.auto_orient = False
-        ## 3d layer flag
-        self.threedimensional = False
-        ## Hidden layer
-        self.hidden = None
-        ## Layer index in AE. Used for parenting and expressions.
-        self.index = None
 
         ## CSS class used by the SVG renderer
         self.css_class = None
         ## `id` attribute used by the SVG renderer
         self.layer_xml_id = None
+        ## tag name used by the SVG renderer
+        self.layer_xml_tag_name = None
+
         ## Whether motion blur is enabled for the layer
         self.motion_blur = None
+        ## Styling effects for this layer
+        self.layer_style = None
 
-        ## In Point of layer. Sets the initial frame of the layer.
-        self.in_point = None
-        ## Out Point of layer. Sets the final frame of the layer.
-        self.out_point = None
-        ## Start Time of layer. Sets the start time of the layer.
-        self.start_time = 0
         ## List of Effects
         self.effects = None
         ## Layer Time Stretching
         self.stretch = 1
-        ## Layer Parent. Uses ind of parent.
-        self.parent_index = None
         ## List of Masks
         self.masks = None
         ## Blend Mode
@@ -134,19 +175,6 @@ class Layer(VisualObject):
             if layer.parent_index == self.index:
                 yield layer
 
-    @classmethod
-    def _load_get_class(cls, lottiedict):
-        if not Layer._classses:
-            Layer._classses = {
-                sc.type: sc
-                for sc in Layer.__subclasses__()
-            }
-        type_id = lottiedict["ty"]
-        if type_id not in Layer._classses:
-            warnings.warn("Unknown layer type: %s" % type_id)
-            return Layer
-        return Layer._classses[type_id]
-
     def __repr__(self):
         return "<%s %s %s>" % (type(self).__name__, self.index, self.name)
 
@@ -163,8 +191,9 @@ class Layer(VisualObject):
         self.composition.remove_layer(self)
 
 
+
 ## @ingroup Lottie
-class NullLayer(Layer):
+class NullLayer(VisualLayer):
     """!
     Layer with no data, useful to group layers together
     """
@@ -172,11 +201,11 @@ class NullLayer(Layer):
     type = 3
 
     def __init__(self):
-        Layer.__init__(self)
+        super().__init__()
 
 
 ## @ingroup Lottie
-class TextLayer(Layer):
+class TextLayer(VisualLayer):
     _props = [
         LottieProp("data", "t", TextAnimatorData, False),
     ]
@@ -184,13 +213,13 @@ class TextLayer(Layer):
     type = 5
 
     def __init__(self):
-        Layer.__init__(self)
+        super().__init__()
         ## Text Data
         self.data = TextAnimatorData()
 
 
 ## @ingroup Lottie
-class ShapeLayer(Layer):
+class ShapeLayer(VisualLayer):
     """!
     Layer containing ShapeElement objects
     """
@@ -201,7 +230,7 @@ class ShapeLayer(Layer):
     type = 4
 
     def __init__(self):
-        Layer.__init__(self)
+        super().__init__()
         ## Shape list of items
         self.shapes = [] # ShapeElement
 
@@ -216,7 +245,7 @@ class ShapeLayer(Layer):
 
 ## @ingroup Lottie
 ## @todo SIF I/O
-class ImageLayer(Layer):
+class ImageLayer(VisualLayer):
     _props = [
         LottieProp("image_id", "refId", str, False),
     ]
@@ -224,13 +253,13 @@ class ImageLayer(Layer):
     type = 2
 
     def __init__(self, image_id=""):
-        Layer.__init__(self)
+        super().__init__()
         ## id pointing to the source image defined on 'assets' object
         self.image_id = image_id
 
 
 ## @ingroup Lottie
-class PreCompLayer(Layer):
+class PreCompLayer(VisualLayer):
     _props = [
         LottieProp("reference_id", "refId", str, False),
         LottieProp("time_remapping", "tm", Value, False),
@@ -241,7 +270,7 @@ class PreCompLayer(Layer):
     type = 0
 
     def __init__(self, reference_id=""):
-        Layer.__init__(self)
+        super().__init__()
         ## id pointing to the source composition defined on 'assets' object
         self.reference_id = reference_id
         ## Comp's Time remapping
@@ -256,7 +285,7 @@ ColorString = LottieValueConverter(color_from_hex, color_to_hex, "Color string")
 
 
 ## @ingroup Lottie
-class SolidColorLayer(Layer):
+class SolidColorLayer(VisualLayer):
     """!
     Layer with a solid color rectangle
     """
@@ -269,10 +298,79 @@ class SolidColorLayer(Layer):
     type = 1
 
     def __init__(self, color=Color(), width=512, height=512):
-        Layer.__init__(self)
+        super().__init__()
         ## Color of the layer as a @c \#rrggbb hex
         self.color = color
         ## Height of the layer.
         self.height = height
         ## Width of the layer.
         self.width = width
+
+
+#ingroup Lottie
+class CameraLayer(Layer):
+    """!
+    3D Camera
+    """
+    type = 13
+    _props = [
+        LottieProp("type", "ty", int, False),
+        LottieProp("transform", "ks", Transform, False),
+        LottieProp("perspective", "pe", Value, False),
+    ]
+
+    def __init__(self):
+        super().__init__()
+
+        ## Layer transform
+        self.transform = None
+        ## Distance from the Z=0 plane.
+        ## Small values yield a higher perspective effect.
+        self.perspective = None
+
+
+#ingroup Lottie
+class DataLayer(Layer):
+    _props = [
+        LottieProp("type", "ty", int, False),
+        LottieProp("data_source_id", "refId", str, False),
+    ]
+    type = 15
+
+    def __init__(self):
+        super().__init__()
+
+        ## ID of the data source in assets
+        self.data_source_id = None
+
+
+#ingroup Lottie
+class AudioSettings(LottieObject):
+    _props = [
+        LottieProp("level", "lv", MultiDimensional, False),
+    ]
+
+    def __init__(self):
+        super().__init__()
+
+        self.level = MultiDimensional([0, 0])
+
+
+#ingroup Lottie
+class AudioLayer(Layer):
+    """!
+    A layer playing sounds
+    """
+    _props = [
+        LottieProp("type", "ty", int, False),
+        LottieProp("audio_settings", "au", AudioSettings, False),
+        LottieProp("sound_id", "refId", str, False),
+    ]
+    type = 6
+
+    def __init__(self):
+        super().__init__()
+
+        self.audio_settings = None
+        ## ID of the sound as specified in the assets
+        self.sound_id = None